fix: panic when re-spawning bar on removed display
Previously, it was possible to attempt to re-spawn a bar on an output
that had already been removed, but that we did not receive an
`OutputEvent::Removed` for yet. This would cause a panic.

This means that if a bar is removed by the compositor for any reason
other than an output removal, we do not attempt to re-spawn it. If this
ever becomes a problem, it's probably possible to fix by waiting a small
duration for any `OutputEvent::Removed` events before re-spawning.